product-service/src/main/java/com/rainhard/product/service/service/ProductService.java
package com.rainhard.product.service.service;

import com.rainhard.product.service.model.Product;
import com.rainhard.product.service.repository.ProductRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class ProductService {

    private final ProductRepository productRepository;

    public Product saveProduct(Product product){
        return productRepository.save(product);
    }

    public List<Product> getAllProduct(){
        return productRepository.findAll();
    }
}
